refactor(lib): update serde and serde_json to 0.9
The Serde update breaks the API, introducing a wrapper around the
previous type alias for a BTreeMap. It also updates `to_value` to
return a Result instead of a Value directly. I've updated the uses
of BTreeMap to refer to the new Map struct instead, and added a
`.expect` to the use of `to_value` - I don't think there's any
reason it should fail, since our data is definitely serializable,
but let me know if you'd prefer a different technique there.